vm: PSB3 container + u32/i32 widths + reserved-range dispatch (VM 2.0 Phase 1a-1d)
Format epoch, hard cutover from PSB2 (Docs/pscal_vm2_plan.md §5.1-5.4):

- Phase 1a: 256-entry computed-goto dispatch table with holes pointing at
  LABEL_INVALID, replacing the branch-before-dispatch bounds check. Verified
  crash-free against 856+ single-byte-flip/truncation mutations of a real
  .bc file.

- Phase 1b: PSB3 container in cache.c. Magic 0x50534233, explicit
  little-endian header/directory/section-body encoding via new
  ByteBuf (write) / Cursor (read) primitives, 8-byte-aligned sections
  (CODE/LINE/CONS/BMAP/PROC/TYPE/META). LINES is now a varint run-length
  table instead of a raw per-code-byte int[]. The PSB2 reader/writer are
  deleted, not kept as a fallback; a stale cache entry just misses the
  magic and recompiles. BMAP carries a builtin-registry fingerprint but
  does not act on it this phase (see cache.c comment: resolving ids at
  write time was found to corrupt interpreter state via premature
  builtin-registry init, surfaced as a Pascal closure regression).

- Phase 1c: CALL's and THREAD_CREATE's address operands widen u16->u32
  (new opcodes.def 'W' spec letter); JUMP/JUMP_IF_FALSE displacements
  widen i16->i32. Every compiler.c emission/patch site updated (goto/label,
  break/continue, if/else, case, loops, short-circuit, ternary, the
  peephole optimizer's relocation pass, finalizeBytecode's CALL backpatch),
  plus every vm.c local/field that narrowed a code address to uint16_t.
  A chunk's code section is no longer capped at 64KB of jump/call targets.

- Phase 1d: all .bc reads go through the bounds-checked Cursor; the
  section directory is validated (in-file, non-overlapping) before any
  section body is parsed.

Zero-diff verified against the pre-change build via Tests/vm_diff_harness.py
(401 units; the only true stdout diff was inherent thread-scheduling
nondeterminism in a par{} test) and the full Tests/run_all_suites.py
(Pascal/CLike/exsh/all scope suites green; remaining two failures are
confirmed pre-existing on the unmodified baseline, unrelated to this change).

+++ src/vm/vm.h
@@ -276,7 +276,7 @@ void vmResetExecutionState(VM* vm); // Reset stack/frames so a VM can be reused
 
 // Main function to interpret a chunk of bytecode
 // Takes a BytecodeChunk that was successfully compiled.
-InterpretResult interpretBytecode(VM* vm, BytecodeChunk* chunk, HashTable* globals, HashTable* const_globals, HashTable* procedures, uint16_t entry);
+InterpretResult interpretBytecode(VM* vm, BytecodeChunk* chunk, HashTable* globals, HashTable* const_globals, HashTable* procedures, uint32_t entry);
 void vmNullifyAliases(VM* vm, uintptr_t disposedAddrValue);
 int vmSpawnCallbackThread(VM* vm, VMThreadCallback callback, void* user_data, VMThreadCleanup cleanup);
 int vmSpawnBuiltinThread(VM* vm, int builtinId, const char* builtinName, int argCount,
